+++ riscv/src/register/mcause.rs
@@ -1,5 +1,7 @@
 //! mcause register
 
+pub use crate::interrupt::Trap;
+
 /// mcause register
 #[derive(Clone, Copy, Debug)]
 pub struct Mcause {
@@ -10,109 +12,6 @@ impl From<usize> for Mcause {
     #[inline]
     fn from(bits: usize) -> Self {
         Self { bits }
-    }
-}
-
-/// Trap Cause
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum Trap {
-    Interrupt(Interrupt),
-    Exception(Exception),
-}
-
-/// Interrupt
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(usize)]
-pub enum Interrupt {
-    SupervisorSoft = 1,
-    MachineSoft = 3,
-    SupervisorTimer = 5,
-    MachineTimer = 7,
-    SupervisorExternal = 9,
-    MachineExternal = 11,
-    Unknown,
-}
-
-/// Exception
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(usize)]
-pub enum Exception {
-    InstructionMisaligned = 0,
-    InstructionFault = 1,
-    IllegalInstruction = 2,
-    Breakpoint = 3,
-    LoadMisaligned = 4,
-    LoadFault = 5,
-    StoreMisaligned = 6,
-    StoreFault = 7,
-    UserEnvCall = 8,
-    SupervisorEnvCall = 9,
-    MachineEnvCall = 11,
-    InstructionPageFault = 12,
-    LoadPageFault = 13,
-    StorePageFault = 15,
-    Unknown,
-}
-
-impl From<usize> for Interrupt {
-    #[inline]
-    fn from(nr: usize) -> Self {
-        match nr {
-            1 => Self::SupervisorSoft,
-            3 => Self::MachineSoft,
-            5 => Self::SupervisorTimer,
-            7 => Self::MachineTimer,
-            9 => Self::SupervisorExternal,
-            11 => Self::MachineExternal,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-impl TryFrom<Interrupt> for usize {
-    type Error = Interrupt;
-
-    #[inline]
-    fn try_from(value: Interrupt) -> Result<Self, Self::Error> {
-        match value {
-            Interrupt::Unknown => Err(Self::Error::Unknown),
-            _ => Ok(value as Self),
-        }
-    }
-}
-
-impl From<usize> for Exception {
-    #[inline]
-    fn from(nr: usize) -> Self {
-        match nr {
-            0 => Self::InstructionMisaligned,
-            1 => Self::InstructionFault,
-            2 => Self::IllegalInstruction,
-            3 => Self::Breakpoint,
-            4 => Self::LoadMisaligned,
-            5 => Self::LoadFault,
-            6 => Self::StoreMisaligned,
-            7 => Self::StoreFault,
-            8 => Self::UserEnvCall,
-            9 => Self::SupervisorEnvCall,
-            11 => Self::MachineEnvCall,
-            12 => Self::InstructionPageFault,
-            13 => Self::LoadPageFault,
-            15 => Self::StorePageFault,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-impl TryFrom<Exception> for usize {
-    type Error = Exception;
-
-    #[inline]
-    fn try_from(value: Exception) -> Result<Self, Self::Error> {
-        match value {
-            Exception::Unknown => Err(Self::Error::Unknown),
-            _ => Ok(value as Self),
-        }
     }
 }
 
@@ -129,13 +28,18 @@ impl Mcause {
         self.bits & !(1 << (usize::BITS as usize - 1))
     }
 
-    /// Trap Cause
+    /// Returns the trap cause represented by this register.
+    ///
+    /// # Note
+    ///
+    /// This method returns a **raw trap cause**, which means that values are represented as `usize`.
+    /// To get a target-specific trap cause, use [`Trap::try_into`] with your target-specific M-Mode trap cause types.
     #[inline]
-    pub fn cause(&self) -> Trap {
+    pub fn cause(&self) -> Trap<usize, usize> {
         if self.is_interrupt() {
-            Trap::Interrupt(Interrupt::from(self.code()))
+            Trap::Interrupt(self.code())
         } else {
-            Trap::Exception(Exception::from(self.code()))
+            Trap::Exception(self.code())
         }
     }